// packages/standards-unece/src/models/IUnLocodeCountrySubdivisionRecord.ts
// Copyright 2026 IOTA Stiftung.
// SPDX-License-Identifier: Apache-2.0.
/**
 * UN/LOCODE Country Subdivision Record information.
 */
export interface IUnLocodeCountrySubdivisionRecord {
	/**
	 * The subdivision code.
	 */
	code: string;

	/**
	 * The subdivision label.
	 */
	label: string;

	/**
	 * The subdivision code.
	 */
	type?: string;
}